chore(reg): Extract the Storyboard extensions
